feat: `cmpeth` package for testing C-chain `Body` RLP decoding
